fix: 상태 상위 컴포넌트로 위임
